ProgAssignments/Assignment4/justin/requirements.txt
ansible==2.9.6
appdirs==1.4.4
attrs==22.1.0
Automat==0.8.0
autopage==0.5.1
Babel==2.10.3
blinker==1.4
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
cliff==4.0.0
cloud-init==22.3.4
cmd2==2.4.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==38.0.3
dbus-python==1.2.16
debtcollector==2.5.0
decorator==5.1.1
devscripts===2.20.2ubuntu2
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docker==6.0.1
dogpile.cache==1.1.8
entrypoints==0.3
filelock==3.8.0
httplib2==0.14.0
hyperlink==19.0.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
incremental==16.10.1
iso8601==1.1.0
Jinja2==2.10.1
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.13.0
keyring==18.0.1
keystoneauth1==5.0.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
msgpack==1.0.4
munch==2.5.0
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.8
numpy==1.23.4
oauthlib==3.1.0
openstacksdk==0.102.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.6.1
oslo.config==9.0.0
oslo.context==5.0.0
oslo.i18n==5.1.0
oslo.log==5.0.0
oslo.serialization==5.0.0
oslo.utils==6.0.0
ovs==2.13.8
packaging==21.3
pandas==1.5.1
pbr==5.11.0
pexpect==4.8.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
prettytable==3.3.0
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pycrypto==2.6.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.18.1
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-cinderclient==9.0.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-glanceclient==4.0.1
python-heatclient==3.0.0
python-keystoneclient==5.0.0
python-neutronclient==8.0.0
python-novaclient==18.0.0
python-openstackclient==6.0.0
python-swiftclient==4.0.1
python-zaqarclient==2.3.0
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-unixsocket==0.2.0
requestsexceptions==1.4.0
rfc3986==2.0.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
sortedcontainers==2.1.0
sos==4.4
ssh-import-id==5.10
stevedore==4.1.0
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==27.11.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
virtualenv==20.16.3
wadllib==1.3.3
warlock==2.0.1
wcwidth==0.2.5
websocket-client==1.4.1
wrapt==1.14.1
zipp==3.8.1
zope.interface==4.7.1
